Reorganized folder structure for pagesize
