+++ ui/src/pages/namespace/Events/History/EventsList.tsx
@@ -1,10 +1,7 @@
 import { Dialog, DialogContent } from "~/design/Dialog";
 import { NoPermissions, NoResult, TableCell, TableRow } from "~/design/Table";
 import { Pagination, PaginationLink } from "~/design/Pagination";
-import {
-  useEventsPageSize,
-  usePageSizeActions,
-} from "~/util/store/pagesizes/pagesize";
+import { useEventsPageSize, usePageSizeActions } from "~/util/store/pagesize";
 
 import { Card } from "~/design/Card";
 import { EventSchemaType } from "~/api/events/schema";
@@ -14,7 +11,7 @@ import { FiltersSchemaType } from "~/api/events/schema/filters";
 import PaginationProvider from "~/components/PaginationProvider";
 import { Radio } from "lucide-react";
 import Row from "./Row";
-import { SelectPageSize } from "./components/SelectPageSize";
+import { SelectPageSize } from "../../../../components/SelectPageSize";
 import SendEvent from "./SendEvent";
 import ViewEvent from "./ViewEvent";
 import { useEvents } from "~/api/events/query/get";
